Fix Windows build erorrs
