+++ internal/osutils/autostart/autostart_lin_mac.go
@@ -1,0 +1,12 @@
+// +build !windows
+
+package autostart
+
+type options struct {
+	launchFileName string
+	iconFileName   string
+	iconFileSource string
+	genericName    string
+	comment        string
+	keywords       string
+}
