+++ internal/osutils/autostart/data_win.go
@@ -1,0 +1,5 @@
+// +build windows
+
+package autostart
+
+var data = map[AppName]options{}
